+++ src/game/game.cc
@@ -61,14 +61,14 @@ void Game::run(bool isSmokeTest) {
   Texture texture;
 #endif
 
-  double previous = static_cast<double>(SDL_GetTicks64());
-  double lag = 0.0;
+  float previous = static_cast<float>(SDL_GetTicks64());
+  float lag = 0.0f;
 
   SDL_Event event;
 
   while (isRunning_) {
-    double current = static_cast<double>(SDL_GetTicks64());
-    double elapsed = current - previous;
+    float current = static_cast<float>(SDL_GetTicks64());
+    float elapsed = current - previous;
     previous = current;
     lag += elapsed;
